+++ tests/testthat/test-Rd-inherit.R
@@ -21,7 +21,7 @@ test_that("no options gives default values", {
 
   expect_equal(
     block$inherit$fields,
-    c("params", "return", "description", "details", "seealso")
+    c("params", "return", "description", "details", "seealso", "sections")
   )
 })
 
@@ -168,6 +168,35 @@ test_that("can inherit details from roxygen topic", {
   expect_equal(out$get_field("details")$values, "C")
 })
 
+
+
+# Inherit sections --------------------------------------------------------
+
+test_that("inherits missing sections", {
+    out <- roc_proc_text(rd_roclet(), "
+    #' A.
+    #' @section A:1
+    #' @section B:1
+    a <- function(x) {}
+
+    #' D
+    #'
+    #' @section A:2
+    #' @inherit a sections
+    b <- function(y) {}
+  ")[[2]]
+
+  section <- out$get_field("section")
+  expect_equal(section$title, c("A", "B"))
+  expect_equal(section$content, c("2", "1"))
+})
+
+test_that("can find section in existing docs", {
+  out <- find_sections(find_topic("base::attach"))
+
+  expect_s3_class(out, "roxy_field_section")
+  expect_equal(out$title, "Good practice")
+})
 
 # Inherit parameters ------------------------------------------------------